Time: 16 ms (42.41%), Space: 7.9 MB (98.21%) - LeetHub

+++ 224-basic-calculator/224-basic-calculator.cpp
@@ -6,7 +6,7 @@ public:
         stack<int>st;
         
         int sign=+1;
-        long long curr_num=0;
+        int curr_num=0;
         int ans=0;
         
         for(int i=0;i<n;i++)
